requirements.txt
# Core dependencies first
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.5.0
patsy>=0.5.3  # Required for statsmodels
statsmodels>=0.13.0

# Core data science and machine learning
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Machine learning - tree models
xgboost>=1.6.0
lightgbm>=3.3.0

# Deep learning
tensorflow>=2.10.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=0.19.0

# Data processing utilities
openpyxl>=3.0.0  # Excel file support
requests>=2.28.0  # API calls (for ICD-11)

# Jupyter and interactive analysis
jupyter>=1.0.0
ipykernel>=6.15.0
ipywidgets>=7.7.0

# Logging and monitoring
tqdm>=4.64.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0